# perf: goleveldb backend to use native range limiter