--- a/tm2/pkg/db/go_level_db.go
+++ b/tm2/pkg/db/go_level_db.go
@@ -209,7 +209,6 @@ type goLevelDBIterator struct {
 	start     []byte
 	end       []byte
 	isReverse bool
-	isInvalid bool
 }
 
 var _ Iterator = (*goLevelDBIterator)(nil)