- iterator to use the native range limiter.
- remove duplicated error checks.
- fix: missing nonNilBytes conversion on the batch jobs.
- add benchmarks to measure iterator performance.
- up-to-date stat properties.
- possibly related to #1013.

<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message was included in the description
- [x] Added references to related issues and PRs
- [x] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated graphs](https://gnoland.github.io/benchmarks), if any. More info [here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>

Benchmarks were written to measure the performance difference. It measures the iteration time of 100 keys in a db with 1 million keys from different offsets in each run.

Here, we are particularly interested in _**Subset iteration**_  test result. (the 4th line, BenchmarkGoLevelDBIteratorSubset)

```
## current version
goos: darwin
noarch: arm64
pkg: github.com/gnolang/gno/tm2/pkg/db
BenchmarkGoLevelDBRandomReadsWrites-10    	  316495	      4077 ns/op	     679 B/op	      13 allocs/op
BenchmarkGoLevelDBIterator-10             	   71209	     14991 ns/op	    5534 B/op	     255 allocs/op
BenchmarkGoLevelDBReverseIterator-10      	   21982	     52273 ns/op	   11675 B/op	     302 allocs/op
BenchmarkGoLevelDBIteratorSubset-10       	   60451	     17119 ns/op	    5737 B/op	     259 allocs/op

## this PR
BenchmarkGoLevelDBRandomReadsWrites-10    	  313550	      4076 ns/op	     679 B/op	      13 allocs/op
BenchmarkGoLevelDBIterator-10             	   73024	     13740 ns/op	    5581 B/op	     256 allocs/op
BenchmarkGoLevelDBReverseIterator-10      	   22754	     51091 ns/op	   11716 B/op	     303 allocs/op
BenchmarkGoLevelDBIteratorSubset-10       	   76792	     13662 ns/op	    5120 B/op	     252 allocs/op
```

```
$ go test .
ok  	github.com/gnolang/gno/tm2/pkg/db
```